Sonar: don't compare "dissimilar" types
This is sort of a false positive, since Sonar is considering
ImmutableSet and HashSet as "dissimilar" when in reality
they both implement Set and the equals comparison
works correctly. However, decided to replace ImmutableSet
with Set.of anyway, since that JDK method creates an
unmodifiable Set.

Sonar java:S5845:
Assertions comparing incompatible types should not be made

Part of #61