Reformat constructors in Consul and CacheConfig
These are (non-public) constructors for use by the builders,
so the lage number of parameters is unavoidable.

Reformat these constructors so that each argument is on
a separate line, mainly for readability.

The Sonar warnings will be resolved in the Sonar UI.
The resolution is "Won't fix" for the reasons stated above.

Part of #74