Enable CatalogITest#shouldGetTaggedAddressesForNode
Enable shouldGetTaggedAddressesForNode and change it
so that it only makes assertions about TaggedAddresses
that are not empty. It was failing because a new node named
"node" is created in other tests; this new node does not
have TaggedAddresses and the Optional is empty. This test
always worked when run by itself, but not when the entire
test class was run. It was just due to Maven test ordering that the
shouldGetTaggedAddressesForNodesLists test always passed,
because if it ran after a test that created the "node" node then
it too would fail. So, make the same change there. Probably
a better way to do this would be to figure out any node that
were created by a test and delete them, but for now this works.

Closes #21