Extract JSON from Java test code into resource file
Since we're still limite to Java 11, we can't use text blocks, so
the next best thing is to extract this gigantic blob of string
concatenation into a .json file in src/test/resources and load
it from the test.

I copied and slightly modified code from kiwi-tests's Fixtures
class since this library does not yet depend on other kiwi projects.