Update Docker image name for Consul to hashicorp/consul:1.15 (#238)
* Update Docker image name for Consul to hashicorp/consul:1.15

This is because of the notice posted to hub.docker.com:

Upcoming in Consul 1.16, we will stop publishing official Dockerhub
images and publish only our Verified Publisher images. Users of Docker
images should pull from hashicorp/consul instead of consul. Verified
Publisher images can be found at
https://hub.docker.com/r/hashicorp/consul.

* Increase max await time in SnapshotClientITest to "fix" test failures

The Java 11 build in GitHub actions is failing on this test in Java 11
(buy not 17 or 20) with the error:

 SnapshotClientITest.shouldBeAbleToSaveAndRestoreSnapshot:50 »
 ConditionTimeout Condition with lambda expression in
 org.kiwiproject.consul.SnapshotClientITest was not fulfilled within
 200 milliseconds.

 Attempting to "fix" by increasing the wait time.