Remove dead/unreachable code from QueryOptions
The null checks in QueryOptions#getNodeMetaQuery and #getTagsQuery
are unreachable because you cannot construct a QueryOptions instance
(ImmutableQueryOptions) that contains null nodeMeta or tag values.
Therefore, there is no way in the #getNodeMetaQuery and #getTagsQuery
methods for the null checks to succeed, which implies the code in
the ternary expression is unreachable.

Closes #151