Fix blacklist expiration behavior in BlacklistingConsulFailoverStrategy (#390)
* Factor duplicate logic from if and else branches in computeNextStage.
* Extract blacklist expiration duration logic from
findTargetNotInBlacklist to new method isPastBlacklistDuration.
* Change isRequestViable to call findTargetNotInBlacklist. Because
findTargetNotInBlacklist will remove a target from the blacklist if past
the expiration, this now ensures a correct result. If the target is not
blacklisted, then the request is viable. If it or any other target is
blacklisted but past the expiration, that target is considered viable
and removed from the blacklist. Only if all targets are currently in
blacklist and none have expired is the request considered not viable.
* Create addToBlacklist method to replace duplicate blacklist.put calls.
* Create isBlacklisted method to check if a target is in blacklist.
* Create isNotBlacklisted method to check if a target is not in
blacklist.
* Rename fromRequest to hostAndPortFromRequest for clarity.
* Rename the blacklistingConsulFailoverStrategy field in the test to
just "strategy" since it is much shorter, and because the test is only
related to the BlacklistingConsulFailoverStrategy.
* Rename the "current" parameter in ConsulFailoverStrategy methods to
"request"
* Add more tests. Move the original tests inside a nested class
ComputeNextStage. The new tests are nested inside their own nested
classes for the various scenarios. I kept the tests for blacklsit
expiration for computeNextStage in a separate nested class,
ComputeNextStageExpiration, mainly b/c they are new and are specifically
for testing the expiration logic. Also added dedicated tests for
isRequest and the (internal) isPastBlacklistDuration method.

Fixes #388