Fix flaky ServiceHealthCache race condition test (#458)
Use a single-thread executor and several CountDownLatch instances to
verify the late listener
was added and has been called. This removes timing of the original raw
Thread and ensures
deterministic notification of both listeners within a maximum time
period using "await" on the
latches instead of using Awaitility.

Note that the two methods added to the Awaiting utility class are not
actually used. I added
them but then ended up not needing them after switching to use the
"await" method on the
latches. But, I am going to keep them around "just in case" I need them
later.