Fix flaky EventClientITest#shouldListEventsByName
Consul does not guarantee any ordering of events, so I'm not sure
why I made that assertion several years ago. Change the test to
fire three additional events with the same name and then
assert that the listEvents(String name) method returns only
events with that name. Do NOT assert on the number of events
as that cannot be guaranteed. The docs on the Consul Event
state that "the gossip layer will make a best-effort to deliver the
event, but there is no guaranteed delivery." In addition, the
docs say regarding event data, "it is not persisted and does not
have a total ordering," which means that "you cannot rely on the
order of message delivery."

Even though there is no guaranteed delivery, in this test the
chance of that is very small.

ref: https://developer.hashicorp.com/consul/commands/event