Replace usage of deprecated URL constructors in Urls (#517)
All URL constructors are deprecated. You are supposed to create a URI
and then convert to URL using toURL(). This changes the Urls utility
class to construct URIs and then convert using toURL(). We also have to
handle additional exceptions that we didn't previously, mainly
URISyntaxException thrown from URI constructors. We still need to handle
MalformedURLException since toURL() throws it.

Additionally, we were using the URL constructor:

URL(String protocol, String host, int port, String file)

This takes a "file" which has a very specific meaning in the URL class.
It includes the path, query, and fragment. There is no equivalent URI
constructor. Instead, the comparable URI constructor accepts the path,
query, and fragment as separate arguments.

To keep the functionality the same, we need to parse the "file" argument
into separate path, query, and fragment values, and then pass those to
the URI constructor. To handle URISyntaxException, the
UncheckedURISyntaxException class from kiwi is copied here and thrown
when there is a URI syntax error. It wraps the checked
URISyntaxException.

Last, added Nullable annotation to the "file" argument in the newUrl
method that has "file" as its last argument.